task3-engineering/requirements.txt
fastapi>=0.100.0
psycopg2-binary>=2.9.0
pytest>=7.0.0
httpx>=0.24.0
uvicorn>=0.20.0
